third_party/WebKit/LayoutTests/fast/dom/shadow/table-border.html

<!DOCTYPE html>
<html>
<body>
<table border>
    <tr><td>A</td><td>B</td></tr>
    <tr id="host1"></tr>
</table>
<table border>
    <tr><td>E</td><td>F</td><td>G</td></tr>
    <tr id="host2"><td>I</td></tr>
</table>
<table border>
    <tr><td>K</td></tr>
    <tr id="host3"></tr>
</table>
<div id="host4">
</div>
<script>
host1.webkitCreateShadowRoot().innerHTML = "<td>C</td><td>D</td>"
host2.webkitCreateShadowRoot().innerHTML = "<td>H</td><content /><td>J</td>";
host3.webkitCreateShadowRoot().innerHTML = "<content select='no-match'><td>L</td></content>";

var td = document.createElement("td");
td.innerHTML = "M";
host4.appendChild(td);
shadow4 = host4.webkitCreateShadowRoot();
shadow4.innerHTML = "<table border><tr id='contentParent'></tr></table>";
shadow4.getElementById("contentParent").appendChild(document.createElement("content"));
</script>
</body>
</html>
