src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/table-border-expected.txt

CONSOLE ERROR: line 19: Uncaught TypeError: host1.createShadowRoot is not a function
